move modal to compositions
