+++ packages/constellation/src/components/Compositions/index.ts
@@ -1,2 +1,3 @@
 export * from './Buttons'
+export * from './Modal'
 export * from './Slider'
